requirements.txt
matplotlib==3.5.3
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.8.1
seaborn==0.12.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.66.2
pycocotools==2.0.6
urllib3==1.26.13
PyYAML==6.0.1
